Cast package totalhits from string to long

+++ NuGetResources/NuGetQueryFeed3.cs
@@ -98,7 +98,7 @@ namespace Microsoft.PackageManagement.NuGetProvider
                 long res = -1;
                 if (root.HasProperty("totalhits"))
                 {
-                    res = root.totalhits;
+                    res = (long)root.totalhits;
                     request.Debug(Resources.Messages.TotalPackagesDiscovered, res);
                 }
                 else
